requirements/local.txt
asgiref==3.2.10
Django==3.0.8
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-model-utils==4.0.0
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
Unipath==1.1
pkg-resources==0.0.0